+++ wcomponents-core/src/main/java/com/github/bordertech/wcomponents/ActionEscape.java
@@ -8,4 +8,20 @@ package com.github.bordertech.wcomponents;
  */
 public class ActionEscape extends Escape {
 
+	/**
+	 * Default constructor.
+	 */
+	public ActionEscape() {
+	}
+
+	/**
+	 * Constructor that allows a message and cause to be provided.
+	 *
+	 * @param message the exception message
+	 * @param cause the original cause
+	 */
+	protected ActionEscape(final String message, final Throwable cause) {
+		super(message, cause);
+	}
+
 }