+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/container/WrongStepServerInterceptor_Test.java
@@ -6,16 +6,13 @@ import com.github.bordertech.wcomponents.Environment;
 import com.github.bordertech.wcomponents.MockWEnvironment;
 import com.github.bordertech.wcomponents.Request;
 import com.github.bordertech.wcomponents.UIContext;
-import com.github.bordertech.wcomponents.UIContextImpl;
 import com.github.bordertech.wcomponents.WApplication;
 import com.github.bordertech.wcomponents.util.Config;
 import com.github.bordertech.wcomponents.util.ConfigurationProperties;
 import com.github.bordertech.wcomponents.util.mock.MockRequest;
 import com.github.bordertech.wcomponents.util.mock.MockResponse;
-import org.junit.Assert;
-import org.apache.commons.configuration.Configuration;
 import org.junit.After;
-import org.junit.Before;
+import org.junit.Assert;
 import org.junit.Test;
 
 /**
@@ -26,99 +23,104 @@ import org.junit.Test;
  */
 public class WrongStepServerInterceptor_Test extends AbstractWComponentTestCase {
 
-	/**
-	 * Backing component.
-	 */
-	private MyBackingComponent component;
-	/**
-	 * Interceptor being tested.
-	 */
-	private WrongStepServerInterceptor interceptor;
-	/**
-	 * User context.
-	 */
-	private UIContext uic;
-	/**
-	 * Mock request.
-	 */
-	private MockRequest request;
-	/**
-	 * Original config.
-	 */
-	private static Configuration originalConfig;
+	@After
+	public void resetConfig() {
+		Config.reset();
+	}
 
-	@Before
-	public void setUp() {
-		component = new MyBackingComponent();
-		interceptor = new WrongStepServerInterceptor();
+	@Test
+	public void testServiceRequestDefaultState() {
+
+		MyBackingComponent component = new MyBackingComponent();
+		WrongStepServerInterceptor interceptor = new WrongStepServerInterceptor();
 		interceptor.setBackingComponent(component);
-		uic = new UIContextImpl();
+		UIContext uic = createUIContext();
 		uic.setUI(component);
 		uic.setEnvironment(new MockWEnvironment());
 
 		setActiveContext(uic);
 
-		request = new MockRequest();
-
 		// Default to GET method
+		MockRequest request = new MockRequest();
 		request.setMethod("GET");
 
-		originalConfig = Config.getInstance();
-	}
-
-	@After
-	public void tearDown() {
-		// Remove overrides
-		Config.setConfiguration(originalConfig);
-	}
-
-	@Test
-	public void testServiceRequestDefaultState() {
 		// Test default state
 		interceptor.serviceRequest(request);
-		Assert.assertTrue("Action phase should have occurred by default",
-				component.handleRequestCalled);
-		Assert.assertEquals("Step should not have been incremented", 0, uic.getEnvironment().
-				getStep());
+		Assert.assertTrue("Action phase should have occurred by default", component.handleRequestCalled);
+		Assert.assertEquals("Step should not have been incremented", 0, uic.getEnvironment().getStep());
 		interceptor.preparePaint(request);
 		Assert.assertEquals("Step should have been incremented", 1, uic.getEnvironment().getStep());
 	}
 
 	@Test
 	public void testServiceRequestCorrectSequence() {
+		MyBackingComponent component = new MyBackingComponent();
+		WrongStepServerInterceptor interceptor = new WrongStepServerInterceptor();
+		interceptor.setBackingComponent(component);
+		UIContext uic = createUIContext();
+		uic.setUI(component);
+		uic.setEnvironment(new MockWEnvironment());
+
+		setActiveContext(uic);
+
+		// Default to GET method
+		MockRequest request = new MockRequest();
+		request.setMethod("GET");
+
 		uic.getEnvironment().setStep(3);
 		request.setParameter(Environment.STEP_VARIABLE, "3");
 
 		interceptor.serviceRequest(request);
-		Assert.assertTrue("Action phase should have occurred for no step error",
-				component.handleRequestCalled);
-		Assert.assertEquals("Step should not have been incremented", 3, uic.getEnvironment().
-				getStep());
+		Assert.assertTrue("Action phase should have occurred for no step error", component.handleRequestCalled);
+		Assert.assertEquals("Step should not have been incremented", 3, uic.getEnvironment().getStep());
 		interceptor.preparePaint(request);
 		Assert.assertEquals("Step should have been incremented", 4, uic.getEnvironment().getStep());
 	}
 
 	@Test
 	public void testServiceRequestIncorrectSequence() {
+
+		MyBackingComponent component = new MyBackingComponent();
+		WrongStepServerInterceptor interceptor = new WrongStepServerInterceptor();
+		interceptor.setBackingComponent(component);
+		UIContext uic = createUIContext();
+		uic.setUI(component);
+		uic.setEnvironment(new MockWEnvironment());
+
+		setActiveContext(uic);
+
+		// Default to GET method
+		MockRequest request = new MockRequest();
+		request.setMethod("GET");
+
 		uic.getEnvironment().setStep(3);
 		request.setParameter(Environment.STEP_VARIABLE, "1");
 		interceptor.serviceRequest(request);
-		Assert.assertTrue("Handle Step Error should have been called for step error",
-				component.handleStepErrorCalled);
-		Assert.assertFalse("Action phase should not have occurred for step error",
-				component.handleRequestCalled);
-		Assert.assertEquals("Step should not have been incremented", 3, uic.getEnvironment().
-				getStep());
+		Assert.assertTrue("Handle Step Error should have been called for step error", component.handleStepErrorCalled);
+		Assert.assertFalse("Action phase should not have occurred for step error", component.handleRequestCalled);
+		Assert.assertEquals("Step should not have been incremented", 3, uic.getEnvironment().getStep());
 		interceptor.preparePaint(request);
 		Assert.assertEquals("Step should have been incremented", 4, uic.getEnvironment().getStep());
 	}
 
 	@Test
 	public void testServiceRequestIncorrectSequenceWithRedirect() {
+
+		MyBackingComponent component = new MyBackingComponent();
+		WrongStepServerInterceptor interceptor = new WrongStepServerInterceptor();
+		interceptor.setBackingComponent(component);
+		UIContext uic = createUIContext();
+		uic.setUI(component);
+		uic.setEnvironment(new MockWEnvironment());
+
+		setActiveContext(uic);
+
+		// Default to GET method
+		MockRequest request = new MockRequest();
+		request.setMethod("GET");
+
 		// Set redirect parameter
-		Configuration config = Config.copyConfiguration(originalConfig);
-		config.addProperty(ConfigurationProperties.STEP_ERROR_URL, "test.url");
-		Config.setConfiguration(config);
+		Config.getInstance().addProperty(ConfigurationProperties.STEP_ERROR_URL, "test.url");
 
 		uic.getEnvironment().setStep(3);
 		request.setParameter(Environment.STEP_VARIABLE, "1");
@@ -128,12 +130,9 @@ public class WrongStepServerInterceptor_Test extends AbstractWComponentTestCase 
 			interceptor.serviceRequest(request);
 			Assert.fail("Interceptor did not detect wrong step");
 		} catch (ActionEscape e) {
-			Assert.assertFalse("Action phase should not have occurred for step error with redirect",
-					component.handleRequestCalled);
-			Assert.assertFalse("Handle Step Error should not have occurred for redirect",
-					component.handleStepErrorCalled);
-			Assert.assertEquals("Step should not have been incremented", 3, uic.getEnvironment().
-					getStep());
+			Assert.assertFalse("Action phase should not have occurred for step error with redirect", component.handleRequestCalled);
+			Assert.assertFalse("Handle Step Error should not have occurred for redirect", component.handleStepErrorCalled);
+			Assert.assertEquals("Step should not have been incremented", 3, uic.getEnvironment().getStep());
 		}
 	}